requirements.txt
attrs==21.4.0
aws-cdk-lib==2.27.0
aws-cdk.region-info==1.159.0
boto3==1.24.30
botocore==1.27.30
cattrs==22.1.0
constructs==10.1.31
exceptiongroup==1.0.0rc7
jmespath==1.0.1
jsii==1.59.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.10
